HSEARCH-2269 Use a delegation pattern in ElasticsearchBridgeProvider
This will ease up the introduction of new bridges and keep everything organized.
